--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -2,7 +2,6 @@ package hw10programoptimization
 
 import (
 	"bufio"
-	"fmt"
 	"io"
 	"strings"
 )
@@ -21,49 +20,37 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
+	return getUsersAndCountDomains(r, domain)
 }
 
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
+func getUsersAndCountDomains(r io.Reader, domain string) (DomainStat, error) {
 	s := bufio.NewScanner(r)
 
-	i := 0
-	for s.Scan() {
-		line := s.Text()
-		user := User{}
-		if err = user.UnmarshalJSON([]byte(line)); err != nil {
-			return
-		}
-		result[i] = user
-		i++
-	}
-
-	if err := s.Err(); err != nil {
-		return result, err
-	}
-
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 	suffix := "." + domain
-	matched := false
 
-	for _, user := range u {
+	var pieces []string
+	var name string
+	var matched bool
+	var user User
+	for s.Scan() {
+		line := s.Bytes()
+		user = User{}
+		if err := user.UnmarshalJSON(line); err != nil {
+			return result, err
+		}
 		matched = strings.HasSuffix(user.Email, suffix)
 		if matched {
-			name := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+			pieces = strings.SplitN(user.Email, "@", 2)
+			name = strings.ToLower(pieces[1])
 			num := result[name]
 			num++
 			result[name] = num
 		}
+	}
+
+	if err := s.Err(); err != nil {
+		return result, err
 	}
 
 	return result, nil